// types.ts
export type Mnemonic =
  | "wallet_waitlist"
  | "jupiter_swap"
  | "wallet_backpack"
  | "bridged_degod"
  | "burn_nft"
  | "joined_telegram"
  | "follow_canydlabs_x"
  | "retardio_casino"
  | "hold_nft_5"
  | "hold_nft_10"
  | "hold_nft_50"
  | "hold_nft_25"
  | "fumbled_bag"
  | "hold_nft_0"
  | "i_minted"
  | "own_sns_domain"
  | "pengu_airdrop"
  | "before_ftx"
  | "deploy_pump"
  | "share_achievement";